Reproducr draft output format (HTML) based on R Markdown and Distill

Creating and publishing reproducible scientific documents in draft (HTML) mode

reproducr_draft(
  toc = FALSE,
  toc_depth = 3,
  toc_float = TRUE,
  fig_width = 7,
  fig_height = 4,
  fig_retina = 5,
  fig_caption = TRUE,
  dev = "ragg_png",
  fig_showtext = TRUE,
  smart = TRUE,
  code_folding = TRUE,
  self_contained = TRUE,
  highlight = "default",
  highlight_downlit = TRUE,
  mathjax = "default",
  extra_dependencies = NULL,
  theme = NULL,
  css = NULL,
  includes = NULL,
  keep_md = FALSE,
  lib_dir = NULL,
  md_extensions = NULL,
  pandoc_args = NULL,
  ...
)

Arguments

toc_float
Float the table of contents to the left when the article is displayed at widths > 1000px. If set to FALSE or the width is less than 1000px the table of contents will be placed above the article body.

fig_retina
globally sets the fig.retina chunk option (defaults to 5)

dev
Graphics device to use for figure output (defaults to 'ragg_png') from the ragg package

fig_showtext
globally sets the fig.showtext chunk option (defaults to TRUE). Use the showtext package to rely on custom fonts for graphs.

smart
Produce typographically correct output, converting straight quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to ellipses.

code_folding
Include code blocks hidden, and allow users to optionally display the code by clicking a "Show code" button just above the output. Pass a character vector to customize the text of the "Show code" button. code_folding is set to TRUE as default chunk option in the reproducr::reproducr_draft.

highlight
Syntax highlighting style. Supported styles include "default", "rstudio", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", "breezedark", and "haddock". Pass NULL to prevent syntax highlighting.

highlight_downlit
Use the downlit package to highlight R code (including providing hyperlinks to function documentation).

Details

reproducr_draft() provides the HTML format

Examples

if (FALSE) {
# devtools::install_github("jschultecloos/reproducr")
library(reproducr)
rmarkdown::render("reproducr-file.Rmd", "reproducr::reproducr_draft")
}