Reproducr manuscript output format based on R Markdown and Bookdown

Creating and publishing reproducible scientific documents in manuscript (PDF) mode

reproducr_manuscript(
  blinded = FALSE,
  toc = FALSE,
  number_sections = FALSE,
  extra_dependencies = NULL,
  fig_width = 7,
  fig_height = 4,
  fig_crop = TRUE,
  fig_align = "center",
  fig_showtext = TRUE,
  dev = "ragg_png",
  dpi = 500,
  highlight = "default",
  colorlinks = TRUE,
  ...
)

Arguments

blinded
TRUE to remove any author information from the title page (defaults to FALSE)

number_sections
Whether to number section headers: if TRUE, figure/table numbers will be of the form X.i, where X is the current first-level section number, and i is an incremental number (the i-th figure/table); if FALSE, figures/tables will be numbered sequentially in the document from 1, 2, ..., and you cannot cross-reference section headers in this case.

fig_showtext
globally sets the fig.showtext chunk option (defaults to TRUE) (defaults to TRUE). Use the showtext package to rely on custom fonts for graphs.

dev
Graphics device to use for figure output (defaults to 'ragg_png') from the ragg package

highlight
syntax highlighting style in your document (defaults to 'tango'). Other available code highlighting options are all options currently supported by rmarkdown::pdf_document: "pygments", "kate", "monochrome", "espresso", "zenburn", "haddock", "breezedark". Pass NULL to prevent syntax highlighting.

colorlinks
should cross-references and links be colored (defaults to TRUE)

...
Other arguments to be passed to reproducr_manuscript

Details

Available metadata variables include:

fontsize

Font size (e.g. 10pt, 11pt, 12pt)

documentclass

LaTeX document class (currently supported: article or scrartcl)

classoption

Option for documentclass (e.g. DIV=calc); may be repeated

mainfont, sansfont, monofont, mathfont

Document fonts (works only with xelatex and lualatex, see the latex_engine option)

linkcolor, urlcolor, citecolor

Color for internal, external, and citation links (red, green, magenta, cyan, blue, black)

linestretch

Options for line spacing (e.g. 1, 1.25, 1.5)

indent

if true, pandoc will use document class settings for indentation (the default LaTeX template otherwise removes indentation and adds space between paragraphs)

reproducr_manuscript() is a PDF output format function that is optimised for reproducible scientific research and is based on R Markdown and Bookdown. Learn more about how to use reproducr_manuscript here: https://jschultecloos.github.io/reproducr/.

References

See https://jschultecloos.github.io/reproducr/ for examples.

Examples

if (FALSE) {
# devtools::install_github("jschultecloos/reproducr")
library(reproducr)
rmarkdown::render("reproducr-file.Rmd", "reproducr::reproducr_manuscript")
}